가속도 측정 로직

+++ wear/src/main/java/com/example/collecthealthdata/domain/local/AccelData.kt
@@ -1,8 +1,8 @@
 package com.example.collecthealthdata.domain.local
 
 data class AccelData(
-    val x: Float,
-    val y: Float,
-    val z: Float,
+    val x: Int,
+    val y: Int,
+    val z: Int,
     val timestamp: Long
 )
